refactoring on editPanels, Added editPanel Folder so EditPanel/types/<differentEditTypes>
