+++ src/components/desktopLayout/editPanel/types/SectionEditor.jsx
@@ -1,70 +1,9 @@
-import React, { useMemo, useState, useEffect } from "react";
+
+import React, { useState, useEffect, useMemo } from "react";
 import OptionListEditor from "./OptionListEditor";
-import fieldTypes from "../../../fields/fieldTypes-config";
-import { updateField, updateChildField, deleteChildField } from "../../../utils/formActions";
-
-export default function EditPanel({
-  isPreview,
-  selectedField,
-  formData,
-  setFormData,
-  onActiveChildChange
-}) {
-
-  if (isPreview) return null;
-
-  const isNone = !selectedField;
-  const isSection = selectedField?.fieldType === "section";
-
-  return (
-    <div className="p-4 bg-white border border-gray-200 rounded-lg shadow-sm mb-2">
-      {/* Placeholder */}
-      {isNone && (
-        <div className="text-gray-600">
-          <h3 className="text-lg font-semibold mb-2">Edit</h3>
-          <p>Select a field in the center panel to edit its properties.</p>
-        </div>
-      )}
-
-      {/* Non-section editor */}
-      {!isNone && !isSection && (
-        <NonSectionEditor f={selectedField} formData={formData} setFormData={setFormData} />
-      )}
-
-      {/* Section editor */}
-      {!isNone && isSection && (
-        <SectionEditor section={selectedField} formData={formData} setFormData={setFormData} onActiveChildChange={onActiveChildChange} />
-      )}
-    </div>
-  );
-}
-
-// Non Section Editor
-function NonSectionEditor({ f, formData, setFormData }) {
-  const onUpdateField = (key, value) => updateField(formData, setFormData, f.id, key, value);
-  const isChoice = ["radio", "check", "selection"].includes(f.fieldType);
-
-  return (
-    <>
-      <h3 className="text-lg font-semibold mb-3">Edit</h3>
-      <CommonEditor f={f} onUpdateField={onUpdateField} />
-
-      {f.fieldType === "input" && (
-        <div className="mt-4">
-          <div className="text-sm font-medium mb-1">Default Answer</div>
-          <input
-            className="w-full px-3 py-2 border border-black/20 rounded"
-            value={f.answer || ""}
-            onChange={(e) => onUpdateField("answer", e.target.value)}
-            placeholder="Default value"
-          />
-        </div>
-      )}
-
-      {isChoice && <OptionListEditor field={f} onUpdateField={onUpdateField} />}
-    </>
-  );
-}
+import CommonEditor from "./CommonEditor";
+import fieldTypes from "../../../../fields/fieldTypes-config";
+import { updateField, updateChildField, deleteChildField } from "../../../../utils/formActions";
 
 // Section Editor with child tabs
 function SectionEditor({ section, formData, setFormData, onActiveChildChange }) {
@@ -172,13 +111,6 @@ function SectionEditor({ section, formData, setFormData, onActiveChildChange }) 
                 {isChoiceChild && (
                   <OptionListEditor field={activeChild} onUpdateField={onUpdateChild} />
                 )}
-
-                <button
-                  onClick={onDeleteChild}
-                  className="mt-4 px-3 py-2 text-sm border border-red-300 text-red-700 rounded hover:bg-red-50"
-                >
-                  Delete this field
-                </button>
               </div>
             )}
           </>
@@ -188,49 +120,4 @@ function SectionEditor({ section, formData, setFormData, onActiveChildChange }) 
   );
 }
 
-// Shared editor for common field properties
-function CommonEditor({ f, onUpdateField }) {
-  return (
-    <div className="space-y-3">
-      <div>
-        <label className="block text-sm mb-1">ID</label>
-        <input
-          className="w-full px-3 py-2 border border-black/20 rounded"
-          value={f.id}
-          onChange={(e) => onUpdateField("id", e.target.value)}
-        />
-      </div>
-
-      <div>
-        <label className="block text-sm mb-1">Label / Question</label>
-        <input
-          className="w-full px-3 py-2 border border-black/20 rounded"
-          value={f.question || ""}
-          onChange={(e) => onUpdateField("question", e.target.value)}
-          placeholder="Enter question text"
-        />
-      </div>
-
-      <div>
-        <label className="inline-flex items-center gap-2 text-sm">
-          <input
-            type="checkbox"
-            checked={!!f.required}
-            onChange={(e) => onUpdateField("required", e.target.checked)}
-          />
-          Required
-        </label>
-      </div>
-
-      <div>
-        <label className="block text-sm mb-1">Sublabel (optional)</label>
-        <textarea
-          className="w-full px-3 py-2 border border-black/20 rounded"
-          value={f.sublabel || ""}
-          onChange={(e) => onUpdateField("sublabel", e.target.value)}
-          placeholder="Helper text / description"
-        />
-      </div>
-    </div>
-  );
-}
+export default SectionEditor;
